requirements.txt
flask
flask-session
gunicorn
python-dotenv
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
langgraph
faiss-cpu